Bug fix. Fixes a bug in caom2-artifact-sync which writes source without accompanying collection name to harvestSkipURI table. Moved HarvestResource to caom2persistence. HarvestResource is now used by caom2-remove, caom2harvester and caom2-artifact-sync. Added 'collection' as a required parameter for artifact-download so that the collection name is available to instantiate HarvestResource.
