+++ caom2-artifact-sync/build.gradle
@@ -15,7 +15,7 @@ sourceCompatibility = 1.7
 
 group = 'org.opencadc'
 
-version = '2.3.19'
+version = '2.3.20'
 
 dependencies {
     compile 'log4j:log4j:1.2.+'
