+++ caom2-remove/build.gradle
@@ -18,7 +18,7 @@ group = 'org.opencadc'
 
 mainClassName = 'ca.nrc.cadc.caom2.remove.Main'
 
-version = '1.0.1'
+version = '1.0.2'
 
 dependencies {
     compile 'log4j:log4j:1.2.+'
